+++ src/collective/collectionfilter/resources/collectionfilter.js
@@ -40,7 +40,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 2 - filter rendered as checkboxes
@@ -55,7 +55,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 3 - filter rendered as dropdowns
@@ -72,7 +72,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
         },
@@ -96,11 +96,18 @@ define([
 
         reloadCollection: function (collectionURL) {
             var cl = new this.contentloader(this.$el, {
-                url: collectionURL,
+                url: collectionURL + '&ajax_load=1',
                 target: '#content-core',
                 content: '#content-core',
                 trigger: 'immediate'
             });
+            // TODO: remove this, once ``contentloader`` handles history
+            // updates itself and adds ajax_load.
+            window.history.replaceState(
+                {path: collectionURL},
+                '',
+                collectionURL
+            );
         }
 
     });