+++ src/collective/collectionfilter/resources/collectionfilter-bundle-compiled.js
@@ -143,8 +143,6 @@ define('collectionfilter',[
     'mockup-patterns-contentloader'
 ], function($, Base, contentloader) {
 
-    'use strict';
-
     var CollectionFilter = Base.extend({
         name: 'collectionfilter',
         trigger: '.pat-collectionfilter',
@@ -168,7 +166,10 @@ define('collectionfilter',[
             $('a.filteritem', this.$el).on('click', function (e) {
                 e.stopPropagation();
                 e.preventDefault();
-                var collectionURL = e.target.closest('a').href;  // strange, it's not the anchor element, but a span.
+                // TODO: nextline: strange, it's not the anchor element itself,
+                // but a span. jQuery's closest also catches the root element
+                // itself, so this shouldn't be a problem.
+                var collectionURL = $(e.target).closest('a').attr('href');
 
                 $(this.trigger).trigger(
                     'collectionfilter:reload',
@@ -178,7 +179,7 @@ define('collectionfilter',[
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 2 - filter rendered as checkboxes
@@ -193,7 +194,7 @@ define('collectionfilter',[
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 3 - filter rendered as dropdowns
@@ -210,7 +211,7 @@ define('collectionfilter',[
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
         },
@@ -234,11 +235,18 @@ define('collectionfilter',[
 
         reloadCollection: function (collectionURL) {
             var cl = new this.contentloader(this.$el, {
-                url: collectionURL,
+                url: collectionURL + '&ajax_load=1',
                 target: '#content-core',
                 content: '#content-core',
                 trigger: 'immediate'
             });
+            // TODO: remove this, once ``contentloader`` handles history
+            // updates itself and adds ajax_load.
+            window.history.replaceState(
+                {path: collectionURL},
+                '',
+                collectionURL
+            );
         }
 
     });
@@ -265,5 +273,5 @@ require([
 
 });
 
-define("/Users/peter/workspace/buildout.coredev/src/collective.collectionfilter/src/collective/collectionfilter/resources/collectionfilter-bundle.js", function(){});
+define("/home/_thet/data/dev/agitator/collectionfilter/plone/src/collective.collectionfilter/src/collective/collectionfilter/resources/collectionfilter-bundle.js", function(){});