+++ src/collective/collectionfilter/resources/collectionfilter.js
@@ -4,8 +4,6 @@ define([
     'mockup-patterns-contentloader'
 ], function($, Base, contentloader) {
 
-    'use strict';
-
     var CollectionFilter = Base.extend({
         name: 'collectionfilter',
         trigger: '.pat-collectionfilter',
@@ -29,7 +27,10 @@ define([
             $('a.filteritem', this.$el).on('click', function (e) {
                 e.stopPropagation();
                 e.preventDefault();
-                var collectionURL = e.target.closest('a').href;  // strange, it's not the anchor element, but a span.
+                // TODO: nextline: strange, it's not the anchor element itself,
+                // but a span. jQuery's closest also catches the root element
+                // itself, so this shouldn't be a problem.
+                var collectionURL = $(e.target).closest('a').attr('href');
 
                 $(this.trigger).trigger(
                     'collectionfilter:reload',
@@ -39,7 +40,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 2 - filter rendered as checkboxes
@@ -54,7 +55,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
             // OPTION 3 - filter rendered as dropdowns
@@ -71,7 +72,7 @@ define([
                     }
                 );
 
-                this.reloadCollection(collectionURL + '&ajax_load=1');
+                this.reloadCollection(collectionURL);
             }.bind(this));
 
         },
@@ -95,11 +96,18 @@ define([
 
         reloadCollection: function (collectionURL) {
             var cl = new this.contentloader(this.$el, {
-                url: collectionURL,
+                url: collectionURL + '&ajax_load=1',
                 target: '#content-core',
                 content: '#content-core',
                 trigger: 'immediate'
             });
+            // TODO: remove this, once ``contentloader`` handles history
+            // updates itself and adds ajax_load.
+            window.history.replaceState(
+                {path: collectionURL},
+                '',
+                collectionURL
+            );
         }
 
     });