+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0 (unreleased)
 ----------------
 
+- Register bundle to depend on ``*`` to avoid weird Select2 initialization error.
+  [thet]
+
 - Add ``input_type`` option to be able to better select the type of input.
   Add ``input_type`` support for dropdowns.
   Remove ``as_input`` attribute and provide upgrade step for it.